backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
openai==1.58.1
chromadb==0.6.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
aiofiles==24.1.0
python-multipart==0.0.20
aiosqlite==0.20.0
